+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * `cargoAudit` now delegates to `mkCargoDerivation` instead of `cargoBuild`
 * `cargoDoc` now delegates to `mkCargoDerivation` instead of `cargoBuild`
 * `cargoFmt` now delegates to `mkCargoDerivation` instead of `cargoBuild`
+* `cargoTarpaulin` now delegates to `mkCargoDerivation` instead of `cargoBuild`
 
 ### Fixed
 * Installing binaries now uses the same version of cargo as was used to build